Split out signal handling routines and shutdown-related routines out of giantic main.c TU to make it more manageable.
