+++ main.c
@@ -133,6 +133,8 @@
 #include "statistics.h"
 #include "core_stats.h"
 #include "pvar.h"
+#include "signals.h"
+#include "shutdown.h"
 #include "poll_types.h"
 #include "net/net_tcp.h"
 #include "net/net_udp.h"
@@ -165,16 +167,13 @@ static int own_pgid = 0; /* whether or not we have our own pgid (and it's ok
 static char* version=OPENSIPS_FULL_VERSION;
 static char* flags=OPENSIPS_COMPILE_FLAGS;
 
-/* last signal received */
-static int sig_flag = 0;
-
 static int user_id = 0;
 static int group_id = 0;
 
 /**
  * Print compile-time constants
  */
-void print_ct_constants(void)
+static void print_ct_constants(void)
 {
 #ifdef ADAPTIVE_WAIT
 	printf("ADAPTIVE_WAIT_LOOPS=%d, ", ADAPTIVE_WAIT_LOOPS);
@@ -187,450 +186,6 @@ void print_ct_constants(void)
 #ifdef VERSIONTYPE
 	printf("%s revision: %s\n", VERSIONTYPE, THISREVISION);
 #endif
-}
-
-/**
- * Clean up on exit. This should be called before exiting.
- * \param show_status set to one to display the mem status
- */
-void cleanup(int show_status)
-{
-	LM_INFO("cleanup\n");
-
-	/*clean-up*/
-
-	/* hack: force-unlock the shared memory lock(s) in case
-	   		 some process crashed and let it locked; this will
-	   		 allow an almost gracious shutdown */
-	if (0
-#if defined F_MALLOC || defined Q_MALLOC
-		|| mem_lock
-#endif
-#ifdef HP_MALLOC
-		|| mem_locks
-#endif
-	) {
-#if defined HP_MALLOC && (defined F_MALLOC || defined Q_MALLOC)
-		if (mem_allocator_shm == MM_HP_MALLOC ||
-		        mem_allocator_shm == MM_HP_MALLOC_DBG) {
-			int i;
-
-			for (i = 0; i < HP_HASH_SIZE; i++)
-				lock_release(&mem_locks[i]);
-		} else {
-			shm_unlock();
-		}
-#elif defined HP_MALLOC
-		int i;
-
-		for (i = 0; i < HP_HASH_SIZE; i++)
-			lock_release(&mem_locks[i]);
-#else
-		shm_unlock();
-#endif
-	}
-
-	handle_ql_shutdown();
-	destroy_modules();
-	udp_destroy();
-	tcp_destroy();
-	destroy_timer();
-	destroy_stats_collector();
-	destroy_script_cb();
-	pv_free_extra_list();
-	tr_free_extra_list();
-	destroy_argv_list();
-	destroy_black_lists();
-	free_route_lists(sroutes); // this is just for testing purposes
-#ifdef PKG_MALLOC
-	if (show_status){
-		LM_GEN1(memdump, "Memory status (pkg):\n");
-		pkg_status();
-	}
-#endif
-	cleanup_log_level();
-
-	if (pt && (0
-#if defined F_MALLOC || defined Q_MALLOC
-		|| mem_lock
-#endif
-#ifdef HP_MALLOC
-		|| mem_locks
-#endif
-		))
-		shm_free(pt);
-	pt=0;
-	if (show_status){
-			LM_GEN1(memdump, "Memory status (shm):\n");
-			shm_status();
-	}
-
-	/* zero all shmem alloc vars that we still use */
-	shm_mem_destroy();
-	if (pid_file) unlink(pid_file);
-	if (pgid_file) unlink(pgid_file);
-}
-
-
-/**
- * Send a signal to all child processes
- * \param signum signal for killing the children
- */
-static void kill_all_children(int signum)
-{
-	int r;
-
-	if (!pt)
-		return;
-
-	for (r = 1; r < counted_max_processes; r++) {
-		if (pt[r].pid == -1 || (pt[r].flags & OSS_PROC_DOING_DUMP))
-			continue;
-
-		/* as the PIDs are filled in by child processes, a 0 PID means
-		 * an un-initalized procees; killing an uninitialized proc is
-		 * very dangerous, so better wait for it to finish its init
-		 * sequence by blocking until the pid is populated */
-		while (pt[r].pid == 0)
-			usleep(1000);
-
-		kill(pt[r].pid, signum);
-	}
-}
-
-
-/**
- * SIGALRM "timeout" handler during the attendant's final cleanup,
- * try to leave a core for future diagnostics.
- */
-static void sig_alarm_abort(int signo)
-{
-	/* LOG is not signal safe, but who cares, we are abort-ing anyway :-) */
-	LM_CRIT("BUG - shutdown timeout triggered, dying...\n");
-	abort();
-}
-
-
-/* RPC function send by main process to all worker processes supporting
- * IPC in order to force a gracefull termination
- */
-static void rpc_process_terminate(int sender_id, void *code)
-{
-	#ifdef PKG_MALLOC
-	LM_GEN1(memdump, "Memory status (pkg):\n");
-	pkg_status();
-	#endif
-
-	/* simply terminate the process */
-	LM_DBG("Process %d exiting with code %d...\n",
-		process_no, (int)(long)code);
-
-	exit( (int)(long)code );
-}
-
-
-/* Implements full shutdown sequence (terminate processes and cleanup)
- * To be called ONLY from MAIN process, not from workers !!!
- */
-static void shutdown_opensips( int status )
-{
-	pid_t  proc;
-	int i, n, p;
-	int chld_status;
-
-	set_osips_state( STATE_TERMINATING );
-
-	/* terminate all processes */
-
-	/* first we try to terminate the processes via the IPC channel */
-	for( i=1,n=0 ; i<counted_max_processes; i++) {
-		/* Depending on the processes status, its PID may be:
-		 *   -1 - process not forked yet
-		 *    0 - process forked but not fully configured by core
-		 *   >0 - process fully running
-		 */
-		if (pt[i].pid!=-1) {
-			/* use IPC (if avaiable) for a graceful termination */
-			if ( IPC_FD_WRITE(i)>0 ) {
-				LM_DBG("Asking process %d [%s] to terminate\n", i, pt[i].desc);
-				if (ipc_send_rpc( i, rpc_process_terminate, (void*)0)<0) {
-					LM_ERR("failed to trigger RPC termination for "
-						"process %d\n", i );
-				}
-			} else {
-				while (pt[i].pid==0) usleep(1000);
-				kill(pt[i].pid, SIGTERM);
-			}
-			n++;
-		}
-	}
-
-	/* now wait for the processes to finish */
-	i = GRACEFUL_SHUTDOWN_TIMEOUT * 100;
-	while( i && n ) {
-		proc = waitpid( -1, &chld_status, WNOHANG);
-		if (proc<=0) {
-			/* no process exited so far, do a small sleep before retry */
-			usleep(10000);
-			i--;
-		} else {
-			if ( (p=get_process_ID_by_PID(proc)) == -1 ) {
-				LM_DBG("unknown child process %d ended. Ignoring\n",proc);
-			} else {
-				LM_INFO("process %d(%d) [%s] terminated, "
-					"still waiting for %d more\n", p, proc, pt[p].desc, n-1);
-				/* mark the child process as terminated / not running */
-				pt[p].pid = -1;
-				status |= chld_status;
-				n--;
-			}
-		}
-	}
-
-	if (i==0 && n!=0) {
-		LM_DBG("force termination for all processes\n");
-		kill_all_children(SIGKILL);
-	}
-
-	/* Only one process is running now. Clean up and return overall status */
-	signal(SIGALRM, sig_alarm_abort);
-	alarm(SHUTDOWN_TIMEOUT - i / 100);
-	cleanup(1);
-	alarm(0);
-	signal(SIGALRM, SIG_IGN);
-
-	dprint("Thank you for running " NAME "\n");
-	exit( status );
-}
-
-
-/**
- * Signal handler for the server.
- */
-void handle_sigs(void)
-{
-	pid_t  chld;
-	int    chld_status,overall_status=0;
-	int    i;
-	int    do_exit;
-
-	switch(sig_flag){
-		case 0: break; /* do nothing*/
-		case SIGPIPE:
-				/* SIGPIPE might be rarely received on use of
-				   exec module; simply ignore it
-				 */
-		case SIGUSR1:
-		case SIGUSR2:
-		case SIGHUP:
-				/* ignoring it*/
-				break;
-		case SIGINT:
-		case SIGTERM:
-			/* we end the program in all these cases */
-			if (sig_flag==SIGINT)
-				LM_DBG("SIGINT received, program terminates\n");
-			else
-				LM_DBG("SIGTERM received, program terminates\n");
-
-			shutdown_opensips( 0/*status*/ );
-			break;
-
-		case SIGCHLD:
-			do_exit = 0;
-			while ((chld=waitpid( -1, &chld_status, WNOHANG ))>0) {
-				/* is it a process we know about? */
-				if ( (i=get_process_ID_by_PID( chld )) == -1 ) {
-					LM_DBG("unknown child process %d ended. Ignoring\n",chld);
-					continue;
-				}
-				if (pt[i].flags & OSS_PROC_SELFEXIT) {
-					LM_NOTICE("process %d/%d did selfexit with "
-						"status %d\n", i, chld,  WTERMSIG(chld_status));
-					reset_process_slot(i);
-					continue;
-				}
-				do_exit = 1;
-				/* process the signal */
-				overall_status |= chld_status;
-				LM_DBG("OpenSIPS exit status = %d\n",overall_status);
-
-				if (WIFEXITED(chld_status))
-					LM_INFO("child process %d exited normally,"
-							" status=%d\n", chld,
-							WEXITSTATUS(chld_status));
-				else if (WIFSIGNALED(chld_status)) {
-					LM_INFO("child process %d exited by a signal"
-							" %d\n", chld, WTERMSIG(chld_status));
-#ifdef WCOREDUMP
-					LM_INFO("core was %sgenerated\n",
-							 WCOREDUMP(chld_status) ?  "" : "not " );
-#endif
-				}else if (WIFSTOPPED(chld_status))
-					LM_INFO("child process %d stopped by a"
-								" signal %d\n", chld,
-								 WSTOPSIG(chld_status));
-
-				/* mark the child process as terminated / not running */
-				pt[i].pid = -1;
-			}
-			if (!do_exit)
-				break;
-			LM_INFO("terminating due to SIGCHLD\n");
-			/* exit */
-			shutdown_opensips( overall_status );
-			break;
-
-		default:
-			LM_CRIT("unhandled signal %d\n", sig_flag);
-	}
-	sig_flag=0;
-}
-
-/* the initial SIGSEGV handler, provided by the OS */
-static struct sigaction sa_sys_segv;
-static char sa_sys_is_valid;
-
-static inline int restore_segv_handler(void)
-{
-	LM_DBG("restoring SIGSEGV handler...\n");
-
-	if (!sa_sys_is_valid)
-		return 1;
-
-	if (sigaction(SIGSEGV, &sa_sys_segv, NULL) < 0) {
-		LM_ERR("failed to restore system SIGSEGV handler\n");
-		return -1;
-	}
-
-	LM_DBG("successfully restored system SIGSEGV handler\n");
-
-	return 0;
-}
-
-/**
- * Exit regulary on a specific signal.
- * This is good for profiling which only works if exited regularly
- * and not by default signal handlers
- * \param signo The signal that should be handled
- */
-static void sig_usr(int signo)
-{
-	int status;
-	pid_t pid;
-	UNUSED(pid);
-
-	if (is_main){
-		if (signo == SIGSEGV) {
-			LM_CRIT("segfault in attendant (starter) process!\n");
-			if (restore_segv_handler() != 0)
-				exit(-1);
-			return;
-		}
-
-		if (sig_flag == 0)
-			sig_flag = signo;
-		else /*  previous sig. not processed yet, ignoring? */
-			return;
-	}else{
-		/* process the important signals */
-		switch(signo){
-			case SIGPIPE:
-			case SIGINT:
-			case SIGUSR1:
-			case SIGUSR2:
-			case SIGHUP:
-					/* ignored*/
-					break;
-			case SIGTERM:
-					/* ignore any SIGTERM if not in shutdown sequance (this 
-					 * is marked by the attendent process) */
-					if (get_osips_state()!=STATE_TERMINATING)
-						return;
-					/* if some shutdown already in progress, ignore this one */
-					if (sig_flag==0) sig_flag=signo;
-					else return;
-					/* do the termination */
-					LM_INFO("signal %d received\n", signo);
-					/* print memory stats for non-main too */
-					#ifdef PKG_MALLOC
-					LM_GEN1(memdump, "Memory status (pkg):\n");
-					pkg_status();
-					#endif
-					exit(0);
-					break;
-			case SIGCHLD:
-					while ( (pid = waitpid(-1, &status, WNOHANG))>0 )
-						LM_DBG("SIGCHLD received from %ld (status=%d),"
-							" ignoring\n", (long)pid,status);
-					break;
-			case SIGSEGV:
-					/* looks like we ate some spicy SIP */
-					LM_CRIT("segfault in process pid: %d, id: %d\n",
-					        pt[process_no].pid, process_no);
-					pt[process_no].flags |= OSS_PROC_DOING_DUMP;
-					if (restore_segv_handler() != 0)
-						exit(-1);
-					pkg_status();
-		}
-	}
-}
-
-
-
-/**
- * Install the signal handlers.
- * \return 0 on success, -1 on error
- */
-int install_sigs(void)
-{
-	struct sigaction act;
-
-	memset(&act, 0, sizeof act);
-
-	act.sa_handler = sig_usr;
-	if (sigaction(SIGSEGV, &act, &sa_sys_segv) < 0) {
-		LM_INFO("failed to install custom SIGSEGV handler -- corefiles must "
-		        "now be written within %d sec to avoid truncation!\n",
-		        GRACEFUL_SHUTDOWN_TIMEOUT);
-	} else {
-		LM_DBG("override SIGSEGV handler: success\n");
-		sa_sys_is_valid = 1;
-	}
-
-	if (signal(SIGINT, sig_usr) == SIG_ERR ) {
-		LM_ERR("no SIGINT signal handler can be installed\n");
-		goto error;
-	}
-	/* if we debug and write to a pipe, we want to exit nicely too */
-	if (signal(SIGPIPE, sig_usr) == SIG_ERR ) {
-		LM_ERR("no SIGINT signal handler can be installed\n");
-		goto error;
-	}
-
-	if (signal(SIGUSR1, sig_usr)  == SIG_ERR ) {
-		LM_ERR("no SIGUSR1 signal handler can be installed\n");
-		goto error;
-	}
-	if (signal(SIGCHLD , sig_usr)  == SIG_ERR ) {
-		LM_ERR("no SIGCHLD signal handler can be installed\n");
-		goto error;
-	}
-	if (signal(SIGTERM , sig_usr)  == SIG_ERR ) {
-		LM_ERR("no SIGTERM signal handler can be installed\n");
-		goto error;
-	}
-	if (signal(SIGHUP , sig_usr)  == SIG_ERR ) {
-		LM_ERR("no SIGHUP signal handler can be installed\n");
-		goto error;
-	}
-	if (signal(SIGUSR2 , sig_usr)  == SIG_ERR ) {
-		LM_ERR("no SIGUSR2 signal handler can be installed\n");
-		goto error;
-	}
-	return 0;
-error:
-	return -1;
 }
